GHOST/X11: setup GLEW and GLEW for EGL
This makes EGL optional and disabled by default but sets up GLEW and GLES
to enable the build on demand. GLEW does not support GLX and EGL at the
same time. Most distributions build GLEW with GLX support, so we have to
use the externally provided GLEW and build with EGL support. This
effectively sets WITH_SYSTEM_GLEW to OFF for all Linux configurations.

Differential Revision: https://developer.blender.org/D12034