GHOST/wayland: use Wayland only when 'BLENDER_WAYLAND' is set
When the X11 and Wayland backends are available, only query Wayland when
the BLENDER_WAYLAND environment variable is set and notify the user if this
failed.
This makes sure that Blender will always use X11. User can then opt-in to
Wayland for testing and can continue using X11 in case of issues.

Differential Revision: D11489